Use better default themeing for highlighting
So far if the config provides no themeing for a threadline component for a given
tag, shell defaults are used, which badly breaks the default theme.
Now the following highlight tags are checked. If no themeing has been provided
for any configured tag, the untagged theme will be used.

Cf issue 25.