Generalize highlighting
Highlighting now works for arbitrary single tags (but be warned: demanding
highlighting for a tag and not providing themeing for it will visually ruin your
default theme). The components that are to be highlighted upon a match are now
directly configurable, so the highlighting level is no longer needed.
The ThreadlineWidget uses the new configuration.

Cf issue #25.