Enable highlighting of date and mailcount components
The default for thread_highlight_components is still 'subject', since that
setting is most discreet and could be a greatest common divisor for all users to
agree on.
This commit does however add exemplary themeing of the date and
mailcount components for the 'flagged' tag. You can try them by enabling the
full array of components in your config file:
    thread_highlight_components = date,mailcount,authors,subject

There is no support for highlighting tags, which are the last component in the
ThreadlineWidget. This is partly due to the fact that the TagWidget comes with
its own themeing capabilities, and we don't wanna override those settings.