Don't fallback to lower priority tag themeing
So far if the config provides no themeing for a threadline component for a given
tag, the following highlight tags are checked. Only if no themeing has been
provided for any configured tag, the untagged theme will be used.
This leads to themes for different highlight tags being mixed if a higher
priority tag doesn't define theming for a component a lower priority tag does.

This commit disables fall-through to lower priority themeing in order to avoid
mixing of themes. Therefore a missing themeing for the top priority highlight
tag immediately triggers fallback to the themeing of the parent option.

Cf issue #25.
Closes #211.