Adjust code to new FallbackConfigParser behaviour
Some of the methods in `AlotConfigParser` and code using the config parser used
`get()` from the `FallbackConfigParser` to determine whether an option existed
by inspecting the return value.
This commit accommodates the fact that since
5d8481e733757a376231594c11e5f2f21ebcf086 the `FallbackConfigParser` doesn't
return `None` on error but raises a `NoOptionError`.

Cf https://github.com/pazz/alot/pull/226#commitcomment-836371 and
https://github.com/pazz/alot/pull/226#issuecomment-3356448.