Handle missing highlighting rules
So far exceptions were raised upon detection of an absent or syntactically invalid
`thread_highlight_rules` option.
This commit changes the behaviour to logging an exception and returning an empty
set of rules as a fallback which will lead to the disabling of all highlighting.

Cf https://github.com/pazz/alot/pull/226#issuecomment-3354066.