Reset logging before configuring
Some unknown piece of code causes the `logging` lib to create a default
configuration, which includes a logging level of `logging.WARNING` and a
`StreamHandler`. The existence of this default configuration prevents our call
to `logging.basicConfig()` to set the configured values, which leads to `debug`
and `info` messages not being logged and all messages being streamed to `stderr`
(which is controlled and gets flushed by `urwid`) instead of being written to
the configured log file.

This commit removes the existing `StreamHandler` before calling
`logging.basicConfig()` which makes the latter install the desired logging
configuration.

Closes #170.