Enable highlighting of tags
So far it was not possible to highlight tags by defining themeing options used
upon match of a query.

This commit enables the themeing of tags for the purpose of highlighting.

The config now may define highlighting for default (not explicitely themed) as
well as themed tags. The priority of themes used is 'themed and highlighted',
'themed', 'default and highlighted', 'default'.

The syntax for specifying highlighted tag themes is ambiguous, though. To avoid
unexpected themeing, highlighting suffix should not match actual tag names.
Cf #209.

Cf issue #25.