Refactor parsing of highlighting rules
So far the highlighting rules were parsed every time a `ThreadlineWidget` was
created.

This commit decouples the parsing from the getter. The rules are parsed only
once upon start up and later on only the parsed dict is returned.

This enables deep syntax checking upon reading the config. Syntax errors raise
exceptions that are handled at the top level. No fallback to default values is
implemented.

Closes #230.