add 'User-Agent' header upon msg construction
The format of this string is either taken from the equally named
envelope-header or the general config option 'user_agent' and defaults to
'alot/%(version)s'. The value of this header is this format string where
keyword '%(version)s' is replaced by the version string of the running
instance.

closees #279