smarter From-header construction for replies
This makes ReplyCommand inspect all To/Cc/Delivered-To values
of the original message and uses a contained address for the
replies From-header if it is listed in one of my accounts.
The order of preference is that of my accounts (order in config file),
the used realname part is taken from the matching account.

cf issue #247