move adding Date headers to save/send commands
This moves the addition of a Date header from Envelope.construct_mail
to the envelope handling commands. This allows us to store
mails locally that contain this timestamp, but send out mails
that do not so that the MTA, that will append its own timestamp,
doesn not create a second entry.
issue #326