explicit abooks
This introduces a new syntax to specify addressbooks
in account sections, which now contain an '[[[abook]]]'
subsection: most important entry is 'type':
the rest of the entries is validated according to this type.
ATM the only working one is 'shellcommand', which makes alot instanciate
a MatchSdtoutAddressbook.
required for this type are the options 'command' and 'regexp'