use new db's for writequeue entries in flush
This moves the creation / destruction of the notmuch.Database object
used to change the index to the code for a single writequeue entry:
Instead of using one db for all entries, we let each one have its own.
This *should* allow the "afterwards" callback allow to retrieve data
from an updated index, that already reflects the changes just made.

issue #317