Be more clever when editing mail template, see discussion in #388
If any headers are to be edited along with the body, then put a
separating empty newline between headertext and body. This is the
default; self.edit_only_body==False. Thus, in the callback function,
the call to parse_template() will have only_body=False. The parsing
regexp then picks up an empty line as separating headerlines from
body. Though, if there is no empty line there, the first
non-header-look line will belong to the body ("\n?" in the regexp).

If no headers are to be edited, we set self.edit_only_body to True.
This is passed on to parse_template(), which then will not do any
regexp parsing, but just let the whole template become the body.