safer globbing in Account.store_mail
we use glob to find the path to a newly stored message file.
This makes us use glob.glob1 instead of glob.glob to make sure the mailbox path
is taken as is - and not interpreted as as regular expression, e.g. if it
contains square brackets.
This further ensures we don't access a possibly empty list out of index

closes #401