add db.utils.call_handler
which gets an email part and interactive? flag
and looks up the handler from mailcap, writes the
decoded payload to a tempfile and calls the handler asyncronously,
returnung a Deferred.

this will later be used to asyncronously call mimehandler to render
the messagewidgets: one adds a callback to the returned deferred
that will update a Text-widgets content.

issue #406