interpret missing %s in mailcap handler correctly
in case the mailcap-specified command string does not contain '%s'
to indicate the path to the tempfile containing the data,
the command should expect the input via sdtin.
This makes the necessary change to commands.thread.OpenAttachmentCommand.

Note that this relies on a not yet implemented parameter 'stdin' of ExternalCommand!
cf issue #458