make ExternalCommand work on command lists
instead of command strings. subprocess.Popen
takes a string list that determines the subcommand to call.
This makes ExternalCommand directly accept string lists to pass on to Popen.
This change obsoletes the 'path' parameter, as callers can now simply
construct the command list accordingly.

OpenAttachmentCommand and EditCommand are updated to match this API change.