add helper split_commandstring
that splits a command string into a list of strings to pass on to subprocess.Popen and the like.

This helper is now used throughout the application instead of calling shlex.split directly as this
is a potential source for errors because shlex is not yet able to properly deal with unicode
bytestrings.