read default bindings from default/default.bindings
This re-introduces a separate file containing the default keybindings
for alot. The values from this file are *always* used as fallbacks
and can be overwritten by bindings from the users config.

To be able to unbind a default binding just bind it to an empty string.
The same will allow buffer-local bindings to be overwritten by global
bindings as described in https://github.com/pazz/alot/issues/403#issuecomment-4434948.

cf issue #487