cleanup input handling in UI
This moves the Command creation/application
to UI.input_filter. This function gets directly triggered by the mainloop.

We hardcode "esc" and "enter" to widgets.globals.ChoiceWidget and CompleteEdit and thus ditch the
abstraction to "cancel" and "select" keypresses.
This obsolets the extra InputWrap widget in UI and
the global SendKeypressCommand and generally makes UI less messy.