fix focus placement after tag command
This makes sure that after a tagging operation, if the thread doesn't match the current search any
more and thus gets removed from the list, the focus is shifted more naturally: If there are
remaining entries below the removed one, focus the next lower.  Otherwise (we are at the end of the
list) focus the new last entry.

Previously, if the last entry was removed, the focus shifted to the beginning of the list.

closes #176