HACK: alot/helper.py: don't convert tabs to spaces
A primary use case for any mail reader is pulling patches off an email
list.  Alot does not support "saving emails" directly, so the print_cmd
function can be used in it's place.  However the email body is sanitized
which corrupts many file types such as diffs.

This hack patch comments out the code which converts tabs to spaces.
This demonstrates a desired feature where the email body is kept
unsanitized.  This change prevents diff and patch files from being
corrupted.

A better solution would be to honor the tabwidth configuration option.
While documented it appears that this config option is never used in the
code.  Note that even if tabwidth was honored in configuration the code
would still need to interpret a value of zero correctly to skip the
tabs-to-space conversion entirely.

The author's python fu is weak and he is in a hurry so commenting out
the offending code block is the quickest way.

Signed-off-by: Mike Turquette <mturquette@linaro.org>