add db.util.mimeparse
This will transform a given email.Message into a tree of
parts: each part-dict contains some meta info, its interpretation as
unicode string (via handlers) and a list `children` that points further
down the tree.
Note that this interpretation is done asyncronously and thus also returns
a Defered.
Later on, this will be responsible for decryption or verification of the mail.