Add body_mimetype to tweak extract_body's defaults
Add a body_mimetype global configuration option to choose the default message
part to return as body in extract_body when types is None. Parts of the
preferred type will be returned if present, all text/* parts will be returned if
none parts of the preferred type exist. The default is "text/html", which should
result in the same behavior as before.