remember and re-interpret command*line* for repeat
... commands. At some point we might want to add a command
history for an undo feature. For this I'd keep a history of
the applied Command objects and in reverse order call its .undo() method.
For this to work we need one Command object for each application.
If however we re-use a Command object using 'repeat' we'd most certainly lose
local information on how to undo the first application..