support multiple recipients in encryption
Multiple recipient-keys can be specified. They are shown in the envelope buffer.
A new command is introduced: 'rmencrypt' do remove a key from the encryption
lsit. Unfortunately pygpgme doesn't implement __cmp__() so we have to either
write something similar on our own or use this solution (index of the key in
encryption list)