read configs lazily in SettingsManager
This causes SettingsManager not to read the default configs upon init
when no config paths are passed on.
It is safe because alot.init.main
triggers reading the config files anyway.

This fixes a problem with building the docs on rfd.org due to them not havin configobj available at
build time.