solarized_dark: improve 256 color approximation
Solarized colors for 256 color palettes are not as closely approximated
as they could be.

256_base0 is set to g66 in Urwid's 256 color palette.  This corresponds
to #a8a8a8 in Xterm.  base0 should ideally be #808080.  This matches g50
exactly, so switch to that grayscale value.

256_base1 is set to g70 in Urwid's 256 color palette.  This corresponds
to #b2b2b2 in Xterm.  base1 should ideally be #8a8a8a.  We can't get
this exactly with Urwid's 256 color palette, but #848484 is a closer
approximation.  Switch to g52 to get this grayscale value.

256_cyan is set to #088 in Urwid's 256 color palette.  This corresponds
to #008787 in Xterm.  cyan should ideally be #00afaf.  This matches #0aa
exactly, so switch to that value.

Signed-off-by: Michael Turquette <mturquette@deferred.io>