postpone binding if candidates with same prefix exist
This ensures that commandlines that are bound to a key
are not fired right away but wait a little (config option 'input_timeout')
in case there are other bindings with this key as prefix.

This allows to bind for instance 'g g' to move up while at the same time
one can bind 'g' to global replies. Without this change, the 'g' binding
would be fired right away.