redraw the UI afer focus [last|first] commands
If one replaces some widget without letting urwid know,
one has to manually trigger it to redraw the screen.
We enforce this to happen after focus_first and focus_last
commands, which might replace widgets in search buffers.

For sake of performance this is *not* done after all other
focus movements.