Move email_as_string function from crypto to helper module
The email_as_string function, and the related RFC3156_canonicalize
function, are now used by the ForwardCommand and are not specific
anymore to the crypto routine. So we move them to the global helper
module.

fix an import removal mistake while moving email_as_string function: StringIO was not only used by email_as_string