Introduce header-based preference in sender select
Introduce a header-based preference in sender selection, as proposed
in #771.

In the examples below, assume the accounts are ordered X-Y-Z in the
config file.

Suppose I'm in a conversation, sending mails from account Y. However,
I add my account X to the "CC:" header. If I do not receive any replies
before sending another mail, I want to send it from account Y again,
9 out of 10 times. The old version would prefer account X, because it
lumps all headers in one list and gets a match on account X first.

Suppose I have aliasing in my mailserver, but send mails from several
of these aliases. They are not "aliases" as such, but accounts which
happen to share a mailbox. However, account X is the account where the
mail is ultimately sent to so will always be present in the
"Delivered-To:" header, even if the sender only knows my address for
account Y, which is in the "To:". Obviously, I want any replies to be
sent from account Y, not account X. This is fixed by prioritizing
matches on the "To:" over "Delivered-To:".

The only case which this patch doesn't fix is when somebody replies
in the first situation, adding *both* addresses to the "To:" header.
In this case, however, I would be fine manually editing the proposed
sender address.