Add option to fix servers with non-standard case-insensitive matching
Some email Servers (some versions of Exchange I can confirm) allow
matching email addresses without regard to case, this is problematic
because alot assumes the standard case sensitive behavior.

This patch addresses this issue by adding a new flag to the accounts
configuration "non_standard_case_insensitive", which defaults to False,
but when set to True will cause alot to match accounts without regard to
case. It does this is a non-intrusive manner using a subclass of unicode
with it's comparison methods overwritten to be case insensitive.

Fixes #847