Fix path in docs generating scripts
The path to the present development version of alot should be prepended to
sys.path in order to shadow other versions of alot installed on the system.
Also it should be a directory from which the alot module is loadable.  That is
the root of the git repository and not one level above this.

This was overlooked in #912.