Replace mutable keyword arguments
There are a number of cases of mutable keyword arguments (list and dict
in this case). Mutable keyword arguments are rather dangerous, since any
mutation of the default value is persistent, which will inevitably lead
to bugs.

For example, imagine this code:

def func(def=[]):
    def.append('foo')
    return def

>>> func()
['foo']
>>> func()
['foo', 'foo']

This is almost certainly not what was intended. This code generally uses
the idiom of setting the default value to None, and then assigning with
or `value = value or []` which will replace value with the empty list
(or dict) when value is falsey, like None or another empty list.