Fix "superflous paren" warnings from pylint
In alot/helper.py this is a fairly obvious change. In alot/buffers.py
I've taken the liberty of replacing a somewhat odd use of multiplication
that relies on the bool class (True and False) implementing an __int__
method. These are used to add the 's' to the end of 'message' if there
are more than one message (thus messages), this is much clearer (and
more correct) when implemented as a ternary instead.