db/message: replace __cmp__ with rich comparisons
This replaces the deprecated __cmp__ function with the more modern
__lt__, __eq__, and __ne__, and then uses functools.total_ordering to
implemented the remaining pieces of the rich comparison protocol
automatically. It also make use of the NotImplemented singleton, which
python uses to signal to rich comparators that this class doesn't no how
to compare to the other class, which results in the complementary method
from the other class being used, and if both return NotImplemented than
a TypeError is generated.

There are two reasons to make this change. First, python 3 only supports
rich comparisons, and second rich comparisons are much faster than using
cmp (and the __cmp__ protocol) because cmp is called at least 3 times
per element, while a rich comparitor is called one per element.

functools.total_ordering is in the standard library from 2.7 and 3.3,
and will implement the rich comparison protocol provided that __eq__ and
at least one of __lt__, __le__, __ge__, or __gt__ are implemented. It
does not (oddly enough) implement __ne__, so I've implemented that
manually.