alot/__main__: use the new validators instead of type=argparse.FileType
This replaces the use of type as a validator with the validator keyword.
This is both cleaner and allows the type keyword to be used for setting
the type (if it ever becomes necessary). It also has the advantage of
not opening and closing the file (thus truncating it) or breaking pipes
(like a fifo).

Fixes #996