Print error when trying to close last buffer without enabling setting
If the last buffer if closed (using 'd' or ':bclose'), by default either
nothing will happen, or the user will be prompted whether he/she wants
to close without saving, and then nothing happens. This can be changed
by setting global.quit_on_last_bclose to True (it defaults to False).

This patch changes that behavior to preempt the prompt and adds a UI
notification that the relevant option is unset, and returns. It will
check the same conditions again after the yield statement since the
number of buffers or the options might have changed (it seems rather
unlikely that the options will change, but it's a tiny amount of code
with minimal performance impact and it's better to be safe).

Fixes #968