widgets/globals: Optimize sorting TagWidgets
The calculation wants to sort first by length, then by string value (so
'z' would come before 'aa'). The calculation does this by deciding that
if one value has a length of 1, and the other doesn't. Currently it does
this by calling min, and max, and calling len 4 times. This is
pretty inefficient, and more complicated that necessary.

Instead, this patch uses 'is not' with booleans to figure out if one and
only one of the lengths is 1.