helper: Use kibi and mibibytes instead of kilo and megabytes
Technically a kilobyte (and it's derivatives like megabytes and
gigabytes) are defined as powers of 1000, while a kibibyte (and it's
derivatives like mibibytes and gibibytes) are defined as powers of 1024.

This patch fixes incorrect language and formatting in the humanize_size
function which defined a kilobyte as 1024.

See this wikipedia article for more information:
https://en.wikipedia.org/wiki/Kibibyte