settings: Add a reload method
This patch does a bit of refactoring to the Settings class for it's
setup, and adds a reload method. The refactoring encapsulates the
handling of NoneType for the read_*config methods, and reworks the
constructor just a bit so that it doesn't rely on calling methods to
create instance variables.

The end result is slightly cleaner, and will be useful for adding a
reload command to alot itself.