Fix message count in statusbar after tagging
When the else branch was hit the count was already updated by
SearchBuffer.rebuild() before the hitcount calculation was done and the
final result was obviously off.  This fixes it by just counting the
messages in the search result directly.

The same code is implicitly already called in the other branch of the if
statement.  So it is hopefully not a big overhead.

Fixes #1098