test: add known broken test for split_commandline
Currently the command line can not contain any quoted semicolons.  The
python shlex module that is used for splitting the command string does
two things with quotes:

1. it interprets them in the sense that everything between quotes is not
   split into tokens
2. it removes them from the generated tokens

To prevent the second we do escape quotes before splitting the command
string with shlex but that implicitly prevents the first point as well.