account: Fix Address comparison to b'' and u''
This uses a try/except because comparing an address to an empty string
should be a fairly uncommon event and try/except will be faster than
`if '@' not in other` in the case where '@' is in other.

This stops alot from crashing if there is no 'From' header in the
email. Outlook generates drafts without a 'From' header.

Fixes #1050