tests: replace test_text_plain_and_other
previously, this created a funky message containing an
"application/octet-stream" part without Content-Diposition=attachment.
The test verified that this part was ignored.

Instead, the mail parser now raises an exception if it failed to find a
mailcap entry for mail parts that are not marked as attachment.
A new test is added here to verify this behaviour.