Add intelligent content summary to thread summary widget
When config thread_content_summary is True, some body text is appended to the
displayed email summary strings. The body text is selected through regex
filters, which exclude things like quoted text. This filtering attempts to be
intelligent, but it probably won't be correct 100% of the time. A statistical
approach would result in better quality, but would take much longer to
implement.