Add support for optional diff and quoting highlighting
Introducing couple of new configurations options in [thread] section:
 * diff_add  - applied to lines starting with +
 * diff_del  - applied to lines starting with -
 * diff_head - applied to lines staring with either "@@" or "diff "
 * nested1 to nested5 - for highlighting nested quoted messages (lines
                        starting with something like "> > >")

Due to how urwid and alot handle displaying text - one widget per line
(or whole message) and styling happening per widget - this change works
only with thread_focus_linewise turned on (the default) and the
highlighting is line-wide.

This is a very rough and ready patch to provide those features as fast
as possible and encourage discussion on a proper implementation.