remember querystring in thread wrapper
Currently, the thread list for the search widget is created by
re-querying the db by thread ids. This loses the "match" flag on
individual messages in a thread, so that, e.g., get_newest_date()
lookest at the newest message in the thread instead of the newest match,
so that dates may look out of order. (The list is sorted by matched
dates.)

Thus, remember the querystring and re-query by thread id and the
original query. This restores dates.