display matched and total in mailcount
Now that thread objects remember the original query, enhance the
mailcount to show the number of matched and total messages in a thread
in the search widget analogous to `notmuch search`.

The original width in the default theme was 5: two parentheses plus up
to three digits. We need to increase this 2+3+3+1=9 now (parentheses,
three digits for each of the counts, slash).