hooks: Add a new type of hook for getting a sending address
I get a good deal of mail generated by an exchange mailing list. There
is nothing in these mails to tell that they are sent to any specific
address, but I need to reply to them with a specific address, which is
not my primary. It also doesn't make sense to me to change my primary
address, since I rarely compose new mail with the address I need to
reply with in this case.

This adds a hook that is called when trying to determine which address
to reply from. This hook will be passed the email and the ui instance,
and is a coroutine (it must either return a deferred or be an
inlineCallbacks decorated callable). The later is required for the hook
to be able to use the UI for non-trivial tasks, such as prompting the
user.