utils: Fix types for add_signature_header
currently _handle_signatures will pass `False~ for the error_msg if
there is no error, but the documentation for add_signature_headers says
it only accepts strings.

Don't do either of these, instead use None if there is no error,
otherwise a string, and updated the assert and documentation to match.