refactor widgets.search.Threadline
This factors out the method _build_part, which is used to create local
Text/Column widgets for each indivitual part of the threadline.

The method is now broken in two: one for text (= subject, authors,..)
parts and one for tags (= Columns of TagWidgets) and both those
functions are moved out of the Threadline class.