db/manager: Limit number of threads loaded for a query
Currently alot will create a process for each query, which fills a pipe
with query objects. This means that for particularly large queries the
pipe can be come particularly full, increasing both the memory load and
slowing down opening queries.

This patch attempts to address this by using a pipe-like object which
limits the number of objects that may be placed in it. This "SizedPipe"
will block once the size is reached, and will unblock as items are taken
off the pipe.