Don't try to add Content-Disposition header to email that already has one
Currently in db.utils.extract_body we try to add a Content-Disposition
header, regardless of whether an email already has one. This isn't
actually legal, and individual mime message may only have one C-D. The
correction is to replace the header with the modified header if it
already exists, and to add a new one only if the message doesn't have
one.

Note that I haven't actually seen a message that hits the path that
would need "add_header", only "replace_header". I have however included
it because it would be a behavioral change to not handle that case.

Fixes #1297