Implement `toggleheader` to allow easy set/unset
There is currently no toggletags equivalent for a header field.
To easily switch certain predefined header fields in the config
file on or off it is useful to have the option of binding

  toggleheader key value

to a certain key, making it for the user easy to switch between
different scenarios in which specific headers are wanted or
unwanted.