Unify set/unset/toggleheader command to single cmd
Following the example of sign/unsign/togglesign the handling of
headers using the set and unset commands has now been unified.
This makes it easier to maintain and avoids the special treatment
of FIXME in two (three with the toggleheader command) functions.

I also renamed `toggleheader` to `toggleset` to be more in line
with `togglesign` and `toggletags`.