Insert dedicated tests for header setting
The tests are written for the envelope `SetCommand`. Previous
tests of this seemed to not exist.

The dedicated tests are written for a variety of test cases that
cover

  - setting of header
  - unsetting of header
  - toggling header

All of them are tested in a variety of scenarios where unneeded
`append` flag is supplied or values that are not required. Here
we follow and assume the hierarchy of evaluation as below:

 - if header exists and should be toggled / unset we do not care
   whether we also obtain value or append
 - if header does not exist we add whatever value we get, even if
   append is set

As `From`, `Cc`, `To` have special treatment in `SetCommand` we
include as well for the `From` header three tests to be explicit
that this special treatment doesn't break the header handling.

Tests could be slimlined in terms of lines, but are tried to be
as close to other test cases in this file as possible.