Show decrypted subject in search results.
Memory hole is currently implemented in Enigmail and message headers can
be protected in the encrypted payload of a message. This commit modifies
the update of a thread's subject by extracting the subject from the
encrypted headers when the most recent message contains protected
headers (v1). This results in the decrypted header being shown in the
search results.

See:

  https://github.com/autocrypt/memoryhole
  https://wiki.gnupg.org/OpenPGPEmailSummit201607/MemoryHole
  https://www.enigmail.net/index.php/en/user-manual/advanced-operations

This is part of #1233. Another part is missing to show decrypted headers
in the message tree.