Enable zsh completion of sender addresses
This parses the config file to find the email addresses.

Because the string interpolation happens at run time and not at load
time of the completion script the python snippet is run on each
completion attempt.  This has the advantage that changes to the config
file are picked up without restarting the shell and the snippet is
simple enough to not impose a real delay in completion.